requirements.txt
pycairo==1.20.1
pillow-8.3.2
numpy==1.21.2
